+++ NEWS.md
@@ -2,6 +2,7 @@ Version 0.14.3-dev
 2013-MM-DD
 
  - Return CORS headers when creating layergroups via GET (#92)
+ - Fix http status on database authentication error (#94)
 
 Version 0.14.2
 2013-11-08